requirements.txt
# requirments.py

opencv-python
torch
torchvision
geopy
requests
Pillow
matplotlib
ultralytics

